% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{betweenness_centrality_cpp}
\alias{betweenness_centrality_cpp}
\title{Compute Betweenness Centrality of Network}
\usage{
betweenness_centrality_cpp(A)
}
\arguments{
\item{A}{A numeric matrix representing the adjacency matrix.}
}
\value{
A numeric vector containing the betweenness centrality scores for each node.
}
\description{
This function calculates the betweenness centrality of each node in a network 
             based on the adjacency matrix. It uses the Floyd-Warshall algorithm 
             implemented in C++ to calculate shortest paths between nodes.
}
\examples{
\dontrun{
    A <- matrix(c(0, 1, 1, 0, 0,
                           1, 0, 1, 1, 0,
                           1, 1, 0, 1, 1,
                           0, 1, 1, 0, 1,
                           0, 0, 1, 1, 0), nrow = 5, byrow = TRUE)
    centrality_scores <- betweenness_centrality_cpp(A)
    print(centrality_scores)
}
}
