% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204144.R
\name{graph_plot}
\alias{graph_plot}
\title{Plot Network Graph without Self-Loops}
\usage{
graph_plot(matrix)
}
\arguments{
\item{matrix}{A symmetric weighted adjacency matrix with no self-loops.}
}
\value{
A network plot showing the relationships between nodes.
}
\description{
This function plots a network graph from an adjacency matrix, 
             ensuring that the matrix is square, symmetric, and free of self-loops.
             It uses edge weights to color the edges and adjusts their thickness.
}
\examples{
\dontrun{
    # Example matrix
    adj_matrix <- matrix(c(0, 1, -1, 1, 0, 0, -1, 0, 0), nrow = 3)
    graph_plot(adj_matrix)
}
}
