% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204144.R
\name{network_balance}
\alias{network_balance}
\title{Measure the Balance of a Network}
\usage{
network_balance(matrix)
}
\arguments{
\item{matrix}{A symmetric weighted adjacency matrix.}
}
\value{
A list containing:
        - `network_info`: The number of nodes, positive edges, and negative edges.
        - `triangle_ratios`: The proportion of each triangle type and overall network balance.
}
\description{
This function calculates the balance of a network based on the number of 
             positive and negative edges in the adjacency matrix. It also computes 
             the distribution of triangle types in the network.
}
\examples{
\dontrun{
    adj_matrix <- matrix(c(0, 1, -1, 1, 0, 0, -1, 0, 0), nrow = 3)
    result <- network_balance(adj_matrix)
    print(result)
}
}
