\name{SA24204144}
\alias{SA24204144}
\docType{package}
\title{
  This package provides several functions for etwork analysis .
}
\description{
  R and Rcpp versions of function are provided in this package (\code{graph_plot} and \code{SVR_test} and \code{network_balance} and \code{degree_centrality_cpp} and \code{degree_centrality_cpp})
}
\author{
Xuerui Du

Maintainer: Xuerui Du <laxiaod@mail.ustc.edu.cn>
}
\keyword{ package }
