\name{SA24204144-package}
\alias{SA24204144-package}
\docType{package}
\title{
  This package provides several functions for benchmarking R and Rcpp.
}
\description{
  R and Rcpp versions of gibbs sampler and vaac function are provided in this package (\code{gibbsR} and \code{vaccR} versus \code{gibbsC} and \code{vaccC})
}
\author{
Xuerui Du

Maintainer: Xuerui Du <laxiaod@mail.ustc.edu.cn>
}
\keyword{ package }
